# swagger.yaml
openapi: 3.0.0
info:
  title: Scientific Publications API
  version: 1.0.0
  description: |
    A comprehensive backend API for analyzing scientific publications with NLP and ML capabilities.
    This system provides endpoints for managing publications, authors, citations, topics, and analytics.
    
    ## Features
    - Publication management with full-text search
    - Author profiles with collaboration networks
    - Citation analysis and relationship mapping
    - Topic modeling and trend analysis
    - Advanced analytics and insights
    - Data import/export capabilities
    - Secure authentication and authorization
    
    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your-jwt-token>`
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse. Limits vary by endpoint type.
  contact:
    name: API Support
    email: support@scientific-publications.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000/api
    description: Development server
paths:
...
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  description: Unique username (3-30 characters)
                password:
                  type: string
                  minLength: 6
                  description: Password (minimum 6 characters)
                first_name:
                  type: string
                  description: First name (optional)
                last_name:
                  type: string
                  description: Last name (optional)
                institution:
                  type: object
                  description: Institution information (optional)
                  properties:
                    name:
                      type: string
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                  refreshToken:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates user and returns JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
                - password
              properties:
                identifier:
                  type: string
                  description: Username or email
                password:
                  type: string
                  description: User password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
                  token:
                    type: string
                  refreshToken:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generates new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags:
        - Authentication
      summary: User logout
      description: Logs out the user and invalidates the session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password
      description: Changes authenticated user's password
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - currentPassword
                - newPassword
              properties:
                currentPassword:
                  type: string
                  description: Current password
                newPassword:
                  type: string
                  minLength: 6
                  description: New password
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/profile:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Returns the profile information of the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Publication Routes
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        full_name:
          type: string
        role:
          type: string
        status:
          type: string
        email_verified:
          type: boolean
        created_at:
          type: string
          format: date-time
    UserProfile:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            first_name:
              type: string
            last_name:
              type: string
            avatar:
              type: string
            bio:
              type: string
            institution:
              $ref: '#/components/schemas/Institution'
            research_interests:
              type: array
              items:
                type: string
            academic_profile:
              $ref: '#/components/schemas/AcademicProfile'
            preferences:
              $ref: '#/components/schemas/UserPreferences'
            statistics:
              $ref: '#/components/schemas/UserStatistics'
            profile_completion:
              type: integer
    Institution:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
        position:
          type: string
        country:
          type: string
    AcademicProfile:
      type: object
      properties:
        google_scholar:
          type: string
        researchgate:
          type: string
        linkedin:
          type: string
        personal_website:
          type: string
    UserPreferences:
      type: object
      properties:
        language:
          type: string
        timezone:
          type: string
        email_notifications:
          type: boolean
        newsletter:
          type: boolean
        privacy:
          type: object
          properties:
            profile_visibility:
              type: string
            show_email:
              type: boolean
    UserStatistics:
      type: object
      properties:
        publications_viewed:
          type: integer
        searches_performed:
          type: integer
        downloads:
          type: integer
        citations_added:
          type: integer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              errors:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
...
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              field:
                type: string